fix: change RunTestArray count param from int32 to uint64 to match size_t
sizeof(array)/sizeof(array[0]) yields size_t; passing it to int32 caused
C4267 narrowing warning which is fatal with /WX. Changed count param and
loop variable to uint64 (the repo's size type) to eliminate the warning.

Agent-Logs-Url: https://github.com/EricLeeFriedman/CopilotChess/sessions/0461de51-3300-4318-af5b-e5ddcc169559

+++ src/tests.h
@@ -17,10 +17,10 @@ struct TestEntry {
 // "PASS: <name>" or "FAIL: <name>" to the debug output, and accumulates
 // results into *passed and *total.  All tests are always executed; nothing
 // short-circuits on failure.
-inline void RunTestArray(const TestEntry* tests, int32 count,
+inline void RunTestArray(const TestEntry* tests, uint64 count,
                          int32* passed, int32* total)
 {
-    for (int32 i = 0; i < count; ++i)
+    for (uint64 i = 0; i < count; ++i)
     {
         bool ok = tests[i].fn();
         OutputDebugStringA(ok ? "PASS: " : "FAIL: ");
